apps/PythonFlask/requirements.txt
flask==3.1.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
